move SelectItemPrefixToken to rewrite package
